# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Beljar",
        "description": "Swagger Sinau",
        "contact": {
            "email": "adrian.rahmandanu2@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/auth/oauth": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User get token",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "grant_type": {
                                        "type": "string"
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "client_id": {
                                        "type": "string"
                                    },
                                    "client_secret": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "grant_type": "password",
                                    "username": "xxx@mail.com",
                                    "password": "12345678",
                                    "client_id": "your client id",
                                    "client_secret": "your client secret"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Failed Validation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get List Users",
                "description": "Returns list user data",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Filter by page",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "0"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit Per Page",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "10"
                        }
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Query by name",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List User",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Failed Validation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Adds a new user - with oneOf examples",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "enum": [
                                            "admin",
                                            "staff",
                                            "reguler"
                                        ]
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Missutsan",
                                    "email": "xxx@mail.com",
                                    "password": "12345678",
                                    "password_confirmation": "12345678",
                                    "role": "xxx"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Failed Validation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/user/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get Detail User",
                "description": "Returns user data",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "uuid": {
                                "summary": "An UUID value.",
                                "value": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail User",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Failed Validation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}